All unnecesary files moved to deprecated folders. They will be removed in master branch. Updated dependencies. Install options added for scripts and launch files.
