+++ iliad_hrsi/package.xml
@@ -18,22 +18,39 @@
   <buildtool_depend>catkin</buildtool_depend>
   <!--  build_depend dynamic_reconfigure  build_depend -->
 
-  <build_depend>roscpp</build_depend>
   <build_depend>rospy</build_depend>
+  <build_depend>bayes_people_tracker</build_depend>
+  <build_depend>geometry_msgs</build_depend>
   <build_depend>orunav_msgs</build_depend>
-  <build_depend>driver_base</build_depend>
-  <!--  build_export_depend  dynamic_reconfigure  /build_export_depend -->
+  <build_depend>numpy</build_depend>
+  <build_depend>std_msgs</build_depend>
+  <build_depend>spencer_tracking_msgs</build_depend>
+  <build_depend>tf</build_depend>
+  <build_depend>math</build_depend>
+  <build_depend>tf2_ros</build_depend>
+  <build_depend>visualization_msgs</build_depend>
+  <build_depend>uuid</build_depend>
+  <build_depend>tf2_geometry_msgs</build_depend>
+  <build_depend>threading</build_depend>
+  <build_depend>qsrlib</build_depend>
+  <build_depend>qsrlib_io</build_depend>
 
-  <build_export_depend>roscpp</build_export_depend>
-  <build_export_depend>rospy</build_export_depend>
-  <exec_depend>driver_base</exec_depend>
-  <exec_depend>orunav_msgs</exec_depend>
-  <exec_depend>rqt_ez_publisher</exec_depend>
-
-  <!-- exec_depend   dynamic_reconfigure   /exec_depend -->
-  <exec_depend>roscpp</exec_depend>
   <exec_depend>rospy</exec_depend>
-
+  <exec_depend>bayes_people_tracker</exec_depend>
+  <exec_depend>geometry_msgs</exec_depend>
+  <exec_depend>orunav_msgs</exec_depend>
+  <exec_depend>numpy</exec_depend>
+  <exec_depend>std_msgs</exec_depend>
+  <exec_depend>spencer_tracking_msgs</exec_depend>
+  <exec_depend>tf</exec_depend>
+  <exec_depend>math</exec_depend>
+  <exec_depend>tf2_ros</exec_depend>
+  <exec_depend>visualization_msgs</exec_depend>
+  <exec_depend>uuid</exec_depend>
+  <exec_depend>tf2_geometry_msgs</exec_depend>
+  <exec_depend>threading</exec_depend>
+  <exec_depend>qsrlib</exec_depend>
+  <exec_depend>qsrlib_io</exec_depend>
 
   <!-- The export tag contains other, unspecified, tags -->
   <export>
