+++ iliad_human_aware_navigation/scripts/deprecated/QSR2constraints.py
